stitcher/requirements.txt
fastapi
uvicorn
opencv-python-headless
python-multipart
Pillow
grpcio
grpcio-tools
